+++ libs/bragi/Cargo.toml
@@ -24,8 +24,17 @@ actix = "0.7"
 serde_qs = "0.4"
 futures = "0.1"
 
-[build-dependencies]
-git-version = "0.2"
-
 [dependencies.mimir]
 path = "../mimir"
+
+[dev-dependencies.docker_wrapper]
+path = "../docker_wrapper"
+
+[dev-dependencies.tools]
+path = "../tools"
+
+[dev-dependencies]
+reqwest = "0.9"
+
+[build-dependencies]
+git-version = "0.2"
